Restructure and clean up whitespace
