+++ ChainedAnimation/Types.swift
@@ -1,0 +1,10 @@
+//
+//  Types.swift
+//  ChainedAnimation
+//
+//  Created by Silvan Dähn on 10/09/2016.
+//  Copyright © 2016 Silvan Dähn. All rights reserved.
+//
+
+public typealias Animation = () -> Void
+public typealias Completion = (Bool) -> Void
